changed default IP, shifted error into coap-endpoint crate, created Dockerfile and Makefile
