+++ coap-endpoint/Makefile
@@ -1,0 +1,4 @@
+CURRENT_DIR:=$(strip $(shell dirname $(realpath $(lastword $(MAKEFILE_LIST)))))
+TOP_DIR := $(CURRENT_DIR)/..
+
+include ../Makefile
