+++ coap-endpoint/src/downstream.rs
@@ -1,13 +1,14 @@
 use std::net::SocketAddr;
 
 //use crate::command::wait_for_command;
+use crate::error::CoapEndpointError;
 use async_trait::async_trait;
 use coap_lite::{CoapRequest, CoapResponse, ResponseType};
 use drogue_cloud_endpoint_common::downstream::PublishOutcome;
 use drogue_cloud_endpoint_common::{
     //commands::Commands,
     downstream::{DownstreamSender, DownstreamSink, Publish},
-    error::{CoapEndpointError, EndpointError},
+    error::EndpointError,
 };
 //use drogue_cloud_service_common::Id;
