+++ endpoint-common/src/error.rs
@@ -1,5 +1,4 @@
 use actix_web::{error::PayloadError, http::StatusCode, HttpResponse, ResponseError};
-use coap_lite::ResponseType;
 use drogue_client::error::ClientError;
 use serde::{Deserialize, Serialize};
 use snafu::Snafu;
@@ -90,32 +89,5 @@ impl From<PayloadError> for HttpEndpointError {
 impl From<EndpointError> for HttpEndpointError {
     fn from(err: EndpointError) -> Self {
         HttpEndpointError(err)
-    }
-}
-
-#[derive(Debug)]
-pub struct CoapEndpointError(pub EndpointError);
-
-impl CoapEndpointError {
-    pub fn status_code(&self) -> ResponseType {
-        match self.0 {
-            EndpointError::InvalidFormat { .. } => ResponseType::BadRequest,
-            EndpointError::InvalidRequest { .. } => ResponseType::BadRequest,
-            EndpointError::ConfigurationError { .. } => ResponseType::InternalServerError,
-            EndpointError::AuthenticationServiceError { .. } => ResponseType::ServiceUnavailable,
-            EndpointError::AuthenticationError { .. } => ResponseType::Forbidden,
-        }
-    }
-}
-
-impl core::fmt::Display for CoapEndpointError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl From<EndpointError> for CoapEndpointError {
-    fn from(err: EndpointError) -> Self {
-        CoapEndpointError(err)
     }
 }
